[Silver III] Title: 주유소, Time: 424 ms, Memory: 41840 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -7,25 +7,25 @@ import java.util.*;
 public class Main {
     private BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
     private StringTokenizer st;
-    private ArrayList<Integer> distanceList = new ArrayList<Integer>();
-    private ArrayList<Integer> priceList = new ArrayList<Integer>();
-    private Iterator<Integer> it;
+    private ArrayList<Long> distanceList = new ArrayList<Long>();
+    private ArrayList<Long> priceList = new ArrayList<Long>();
+    private Iterator<Long> it;
     private long sum = 0;
 
     public  void run() throws Exception {
         int count = Integer.parseInt(br.readLine());
         st = new StringTokenizer(br.readLine());
         for(int i=0; i<count -1; i++)
-            distanceList.add(Integer.parseInt(st.nextToken()));
+            distanceList.add(Long.parseLong(st.nextToken()));
 
         st = new StringTokenizer(br.readLine());
         for(int i=0; i<count; i++)
-            priceList.add(Integer.parseInt(st.nextToken()));
+            priceList.add(Long.parseLong(st.nextToken()));
 
-        int minPrice = priceList.get(0);
+        Long minPrice = priceList.get(0);
         for(int i=0; i<count-1; i++){
-            int price = priceList.get(i);
-            int distance = distanceList.get(i);
+            Long price = priceList.get(i);
+            Long distance = distanceList.get(i);
             if(minPrice > price)
                 minPrice = price;
             sum+= minPrice * distance;
